+++ src/crypto/salsa20.rs
@@ -9,12 +9,13 @@ use ffi::{crypto_stream_salsa20,
           crypto_stream_salsa20_NONCEBYTES};
 use libc::c_ulonglong;
 use std::intrinsics::volatile_set_memory;
+use std::iter::repeat;
 use randombytes::randombytes_into;
 
 stream_module!(crypto_stream_salsa20,
                crypto_stream_salsa20_xor,
-               crypto_stream_salsa20_KEYBYTES as uint,
-               crypto_stream_salsa20_NONCEBYTES as uint);
+               crypto_stream_salsa20_KEYBYTES as usize,
+               crypto_stream_salsa20_NONCEBYTES as usize);
 
 #[test]
 fn test_vector_1() {
