+++ src/crypto/salsa208.rs
@@ -9,9 +9,10 @@ use ffi::{crypto_stream_salsa208,
           crypto_stream_salsa208_NONCEBYTES};
 use libc::c_ulonglong;
 use std::intrinsics::volatile_set_memory;
+use std::iter::repeat;
 use randombytes::randombytes_into;
 
 stream_module!(crypto_stream_salsa208,
                crypto_stream_salsa208_xor,
-               crypto_stream_salsa208_KEYBYTES as uint,
-               crypto_stream_salsa208_NONCEBYTES as uint);
+               crypto_stream_salsa208_KEYBYTES as usize,
+               crypto_stream_salsa208_NONCEBYTES as usize);
